Refactor the code to fetch the API.
