--- constants.go
@@ -1,3 +1,0 @@
-package main
-
-const baseURL = "https://data.moviebuff.com"
